requirements.txt
Flask==2.3.3
python-telegram-bot==20.4
openai==1.3.0
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0